fix: @withbee/utils 디렉토리 정리
