+++ apps/mobile/package.json
@@ -13,6 +13,7 @@
     "@withbee/styles": "workspace:*",
     "@withbee/ui": "workspace:*",
     "@withbee/hooks": "workspace:*",
+    "@withbee/utils": "workspace:*",
     "chart.js": "^4.4.6",
     "dayjs": "^1.11.13",
     "next": "14.2.6",
