+++ packages/utils/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@withbee/utils",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "lint": "eslint . --max-warnings 0"
+  },
+  "devDependencies": {
+    "@types/eslint": "^8.56.5",
+    "@types/node": "^20.11.24",
+    "@types/react": "^18.2.61",
+    "@types/react-dom": "^18.2.19",
+    "@withbee/eslint-config": "workspace:*",
+    "@withbee/types": "workspace:*",
+    "@withbee/typescript-config": "workspace:*",
+    "eslint": "^8.57.0",
+    "typescript": "5.5.4"
+  },
+  "dependencies": {
+    "next": "14.2.6",
+    "react": "^18.2.0"
+  }
+}
