+++ packages/utils/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@withbee/typescript-config/react-library.json",
+  "compilerOptions": {
+    "module": "ESNext",
+    "moduleResolution": "nodenext",
+    "target": "ESNext",
+    "outDir": "dist"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
