+++ packages/utils/tsconfig.lint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@withbee/typescript-config/react-library.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src", "turbo"],
+  "exclude": ["node_modules", "dist"]
+}
